backend/RestAPI/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
grpcio 
grpcio-tools
Flask==3.1.1
psycopg2-binary==2.9.10
flask-cors